create MAGES folder and move LAY, MPK into it
